+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to `Laragram` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.4.2] (2022-11-24)
+
+### Added
+
+- Added customatization auth in `config.php` and models
+
+### Changed
+
+- Replaced `language` column on table `laragram_user` to `settings`
+- Replaced type `contains` column on table `laragram_sessions`
+- Improved code
+
 ## [v1.4.1] (2022-11-22)
 
 ### Added